+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Fixed the Localization String for Animal Hybrids (#144) ("Baseliner Bug" still persists, thats all RW not me)
 - Added another null-check for "Offset Lifeforce" (#120,#143)
 - Added a check for LifeForce-Empath having an active life-force gene (#143), which likely threw errors being active despite children not having life-force yet.
-
+- Added a map-check for LifeForce Empaths. Won't work off-map. (#120)
 
 # 2.2.0 (23-07-2024)